+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 11.0.1
 -----------------
 - Fix PHP deprecated passing null to parameter #1 ($datetime) of type string in Warehouse.php
 - MySQL fix infinite loop, emulate PostgreSQL's nextval() in database.inc.php
+- Fix Apache 414 Request-URI Too Long, use POST method instead of GET in Administration.php
 
 Changes in 11.0
 ---------------